HBP-41 Review plot utils
